+++ src/epd/GxEPD2_102.h
@@ -33,7 +33,7 @@ class GxEPD2_102 : public GxEPD2_EPD
     static const uint16_t full_refresh_time = 3600; // ms, e.g. 3507492us
     static const uint16_t partial_refresh_time = 1200; // ms, e.g. 1102149us
     // constructor
-    GxEPD2_102(int8_t cs, int8_t dc, int8_t rst, int8_t busy);
+    GxEPD2_102(int16_t cs, int16_t dc, int16_t rst, int16_t busy);
     // methods (virtual)
     //  Support for Bitmaps (Sprites) to Controller Buffer and to Screen
     void clearScreen(uint8_t value = 0xFF); // init controller memory and screen (default white)
