+++ src/epd/GxEPD2_154_M10.h
@@ -31,7 +31,7 @@ class GxEPD2_154_M10 : public GxEPD2_EPD
     static const uint16_t full_refresh_time = 4000; // ms, e.g. 3665378us
     static const uint16_t partial_refresh_time = 1700; // ms, e.g. 1635335us
     // constructor
-    GxEPD2_154_M10(int8_t cs, int8_t dc, int8_t rst, int8_t busy);
+    GxEPD2_154_M10(int16_t cs, int16_t dc, int16_t rst, int16_t busy);
     // methods (virtual)
     //  Support for Bitmaps (Sprites) to Controller Buffer and to Screen
     void clearScreen(uint8_t value = 0xFF); // init controller memory and screen (default white)
