+++ src/epd/GxEPD2_213_B72.h
@@ -30,7 +30,7 @@ class GxEPD2_213_B72 : public GxEPD2_EPD
     static const uint16_t full_refresh_time = 1700; // ms, e.g. 1686008us
     static const uint16_t partial_refresh_time = 200; // ms, e.g. 192385us
     // constructor
-    GxEPD2_213_B72(int8_t cs, int8_t dc, int8_t rst, int8_t busy);
+    GxEPD2_213_B72(int16_t cs, int16_t dc, int16_t rst, int16_t busy);
     // methods (virtual)
     //  Support for Bitmaps (Sprites) to Controller Buffer and to Screen
     void clearScreen(uint8_t value = 0xFF); // init controller memory and screen (default white)
