+++ src/epd/GxEPD2_290.h
@@ -30,7 +30,7 @@ class GxEPD2_290 : public GxEPD2_EPD
     static const uint16_t full_refresh_time = 1600; // ms, e.g. 1575016us
     static const uint16_t partial_refresh_time = 420; // ms, e.g. 412493us
     // constructor
-    GxEPD2_290(int8_t cs, int8_t dc, int8_t rst, int8_t busy);
+    GxEPD2_290(int16_t cs, int16_t dc, int16_t rst, int16_t busy);
     // methods (virtual)
     //  Support for Bitmaps (Sprites) to Controller Buffer and to Screen
     void clearScreen(uint8_t value = 0xFF); // init controller memory and screen (default white)
