+++ src/epd3c/GxEPD2_583c.h
@@ -30,7 +30,7 @@ class GxEPD2_583c : public GxEPD2_EPD
     static const uint16_t full_refresh_time = 32000; // ms, e.g. 29165492us
     static const uint16_t partial_refresh_time = 32000; // ms, e.g. 29165492us
     // constructor
-    GxEPD2_583c(int8_t cs, int8_t dc, int8_t rst, int8_t busy);
+    GxEPD2_583c(int16_t cs, int16_t dc, int16_t rst, int16_t busy);
     // methods (virtual)
     //  Support for Bitmaps (Sprites) to Controller Buffer and to Screen
     void clearScreen(uint8_t value = 0x33); // init controller memory and screen (default white)
